Move attack goals to new package
