+++ src/main/java/com/mraof/minestuck/entity/AttackingAnimatedEntity.java
@@ -1,7 +1,6 @@
 package com.mraof.minestuck.entity;
 
-import com.mraof.minestuck.entity.ai.MoveToTargetGoal;
-import com.mraof.minestuck.entity.ai.SlowAttackWhenInRangeGoal;
+import com.mraof.minestuck.entity.ai.attack.AttackState;
 import net.minecraft.entity.CreatureEntity;
 import net.minecraft.entity.EntityType;
 import net.minecraft.network.datasync.DataParameter;
@@ -12,7 +11,7 @@ import net.minecraft.world.World;
 /**
  * A base class for animated entities with a potentially delayed attack.
  */
-public abstract class AttackingAnimatedEntity extends CreatureEntity implements SlowAttackWhenInRangeGoal.AttackStateHolder
+public abstract class AttackingAnimatedEntity extends CreatureEntity implements AttackState.Holder
 {
 	private static final DataParameter<Integer> CURRENT_ACTION = EntityDataManager.defineId(AttackingAnimatedEntity.class, DataSerializers.INT);
 	
@@ -29,13 +28,13 @@ public abstract class AttackingAnimatedEntity extends CreatureEntity implements 
 	}
 	
 	@Override
-	public SlowAttackWhenInRangeGoal.AttackState getAttackState()
+	public AttackState getAttackState()
 	{
-		return SlowAttackWhenInRangeGoal.AttackState.values()[this.entityData.get(CURRENT_ACTION)];
+		return AttackState.values()[this.entityData.get(CURRENT_ACTION)];
 	}
 	
 	@Override
-	public void setAttackState(SlowAttackWhenInRangeGoal.AttackState state)
+	public void setAttackState(AttackState state)
 	{
 		this.entityData.set(CURRENT_ACTION, state.ordinal());
 	}
