+++ src/main/java/com/mraof/minestuck/entity/carapacian/PawnEntity.java
@@ -1,7 +1,8 @@
 package com.mraof.minestuck.entity.carapacian;
 
-import com.mraof.minestuck.entity.ai.MoveToTargetGoal;
-import com.mraof.minestuck.entity.ai.SlowAttackWhenInRangeGoal;
+import com.mraof.minestuck.entity.ai.attack.MoveToTargetGoal;
+import com.mraof.minestuck.entity.ai.attack.SlowAttackWhenInRangeGoal;
+import com.mraof.minestuck.entity.ai.attack.ZeroMovementDuringAttack;
 import com.mraof.minestuck.item.MSItems;
 import com.mraof.minestuck.util.AnimationUtil;
 import net.minecraft.enchantment.EnchantmentHelper;
@@ -73,7 +74,7 @@ public class PawnEntity extends CarapacianEntity implements IRangedAttackMob, IM
 	{
 		super.registerGoals();
 		this.goalSelector.addGoal(2, new SlowAttackWhenInRangeGoal<>(this, 6, 12));
-		this.goalSelector.addGoal(2, new SlowAttackWhenInRangeGoal.ZeroMovementDuringAttack<>(this));
+		this.goalSelector.addGoal(2, new ZeroMovementDuringAttack<>(this));
 		this.goalSelector.addGoal(3, new MoveToTargetGoal(this, 1F, false));
 		this.targetSelector.addGoal(2, new NearestAttackableTargetGoal<>(this, LivingEntity.class, 0, true, false, entity -> attackEntitySelector.isEntityApplicable(entity)));
 	}
